Detach FeedType and Feed api calls into FeedsMixin
