+++ gvm/protocols/gmpv208/__init__.py
@@ -60,6 +60,11 @@ from gvm.protocols.gmpv208.entities.entities import (
     EntityType,
     get_entity_type_from_string,
 )
+from gvm.protocols.gmpv208.entities.feeds import (
+    FeedType,
+    FeedsMixin,
+    get_feed_type_from_string,
+)
 from gvm.protocols.gmpv208.entities.hosts import (
     HostsMixin,
 )
@@ -127,6 +132,7 @@ class Gmp(
     AlertsMixin,
     AuditsMixin,
     CredentialsMixin,
+    FeedsMixin,
     HostsMixin,
     NotesMixin,
     OperatingSystemsMixin,
