+++ gvm/protocols/gmpv208/entities/feeds.py
@@ -1,0 +1,92 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2021 Greenbone Networks GmbH
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+# pylint:  disable=redefined-builtin
+# MAYBE we should change filter to filter_string (everywhere)
+
+from enum import Enum
+from typing import Any, Optional
+
+from gvm.errors import InvalidArgument, InvalidArgumentType, RequiredArgument
+from gvm.xml import XmlCommand
+
+
+class FeedType(Enum):
+    """Enum for feed types"""
+
+    NVT = "NVT"
+    CERT = "CERT"
+    SCAP = "SCAP"
+    GVMD_DATA = "GVMD_DATA"
+
+
+def get_feed_type_from_string(feed_type: Optional[str]) -> Optional[FeedType]:
+    """Convert a feed type string into a FeedType instance"""
+    if not feed_type:
+        return None
+
+    try:
+        return FeedType[feed_type.upper()]
+    except KeyError:
+        raise InvalidArgument(
+            argument='feed_type', function=get_feed_type_from_string.__name__
+        ) from None
+
+
+class FeedsMixin:
+    def get_feeds(self) -> Any:
+        """Request the list of feeds
+
+        Returns:
+            The response. See :py:meth:`send_command` for details.
+        """
+        return self._send_xml_command(XmlCommand("get_feeds"))
+
+    def get_feed(self, feed_type: Optional[FeedType]) -> Any:
+        """Request a single feed
+
+        Arguments:
+            feed_type: Type of single feed to get: NVT, CERT or SCAP
+
+        Returns:
+            The response. See :py:meth:`send_command` for details.
+        """
+        if not feed_type:
+            raise RequiredArgument(
+                function=self.get_feed.__name__, argument='feed_type'
+            )
+
+        if not isinstance(feed_type, FeedType):
+            raise InvalidArgumentType(
+                function=self.get_feed.__name__,
+                argument='feed_type',
+                arg_type=FeedType.__name__,
+            )
+
+        cmd = XmlCommand("get_feeds")
+        cmd.set_attribute("type", feed_type.value)
+
+        return self._send_xml_command(cmd)
+
+    def sync_feed(self) -> Any:
+        """Request a synchronization with the NVT feed service
+
+        Returns:
+            The response. See :py:meth:`send_command` for details.
+        """
+        return self._send_xml_command(XmlCommand("sync_feed"))
